requirements.txt
libarchive-c
pyalpm
yaml
